# feat: add clonevt methods