--- a/input/otlp/traces.go
+++ b/input/otlp/traces.go
@@ -149,7 +149,7 @@ func (c *Consumer) convertSpan(
 	name := otelSpan.Name()
 	spanID := hexSpanID(otelSpan.SpanID())
 	representativeCount := getRepresentativeCountFromTracestateHeader(otelSpan.TraceState().AsRaw())
-	event := proto.Clone(baseEvent).(*modelpb.APMEvent)
+	event := baseEvent.CloneVT()
 	initEventLabels(event)
 	event.Timestamp = timestamppb.New(startTime.Add(timeDelta))
 	if id := hexTraceID(otelSpan.TraceID()); id != "" {
@@ -193,7 +193,7 @@ func (c *Consumer) convertSpan(
 	*out = append(*out, event)
 
 	events := otelSpan.Events()
-	event = proto.Clone(event).(*modelpb.APMEvent)
+	event = event.CloneVT()
 	event.Labels = baseEvent.Labels               // only copy common labels to span events
 	event.NumericLabels = baseEvent.NumericLabels // only copy common labels to span events
 	event.Event = nil                             // don't copy event.* to span events
@@ -872,7 +872,7 @@ func (c *Consumer) convertSpanEvent(
 	parent *modelpb.APMEvent, // either span or transaction
 	timeDelta time.Duration,
 ) *modelpb.APMEvent {
-	event := proto.Clone(parent).(*modelpb.APMEvent)
+	event := parent.CloneVT()
 	initEventLabels(event)
 	event.Transaction = nil // populate fields as required from parent
 	event.Span = nil        // populate fields as required from parent