--- a/model/modelpb/datastream_vtproto.pb.go
+++ b/model/modelpb/datastream_vtproto.pb.go
@@ -25,6 +25,7 @@ import (
 	fmt "fmt"
 	io "io"
 
+	proto "google.golang.org/protobuf/proto"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 )
 
@@ -34,6 +35,26 @@ const (
 	// Verify that runtime/protoimpl is sufficiently up-to-date.
 	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
 )
+
+func (m *DataStream) CloneVT() *DataStream {
+	if m == nil {
+		return (*DataStream)(nil)
+	}
+	r := &DataStream{
+		Type:      m.Type,
+		Dataset:   m.Dataset,
+		Namespace: m.Namespace,
+	}
+	if len(m.unknownFields) > 0 {
+		r.unknownFields = make([]byte, len(m.unknownFields))
+		copy(r.unknownFields, m.unknownFields)
+	}
+	return r
+}
+
+func (m *DataStream) CloneMessageVT() proto.Message {
+	return m.CloneVT()
+}
 
 func (m *DataStream) MarshalVT() (dAtA []byte, err error) {
 	if m == nil {