--- a/model/modelpb/destination_vtproto.pb.go
+++ b/model/modelpb/destination_vtproto.pb.go
@@ -25,6 +25,7 @@ import (
 	fmt "fmt"
 	io "io"
 
+	proto "google.golang.org/protobuf/proto"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 )
 
@@ -34,6 +35,25 @@ const (
 	// Verify that runtime/protoimpl is sufficiently up-to-date.
 	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
 )
+
+func (m *Destination) CloneVT() *Destination {
+	if m == nil {
+		return (*Destination)(nil)
+	}
+	r := &Destination{
+		Address: m.Address,
+		Port:    m.Port,
+	}
+	if len(m.unknownFields) > 0 {
+		r.unknownFields = make([]byte, len(m.unknownFields))
+		copy(r.unknownFields, m.unknownFields)
+	}
+	return r
+}
+
+func (m *Destination) CloneMessageVT() proto.Message {
+	return m.CloneVT()
+}
 
 func (m *Destination) MarshalVT() (dAtA []byte, err error) {
 	if m == nil {