--- a/model/modelpb/device_vtproto.pb.go
+++ b/model/modelpb/device_vtproto.pb.go
@@ -25,6 +25,7 @@ import (
 	fmt "fmt"
 	io "io"
 
+	proto "google.golang.org/protobuf/proto"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 )
 
@@ -34,6 +35,45 @@ const (
 	// Verify that runtime/protoimpl is sufficiently up-to-date.
 	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
 )
+
+func (m *Device) CloneVT() *Device {
+	if m == nil {
+		return (*Device)(nil)
+	}
+	r := &Device{
+		Id:           m.Id,
+		Model:        m.Model.CloneVT(),
+		Manufacturer: m.Manufacturer,
+	}
+	if len(m.unknownFields) > 0 {
+		r.unknownFields = make([]byte, len(m.unknownFields))
+		copy(r.unknownFields, m.unknownFields)
+	}
+	return r
+}
+
+func (m *Device) CloneMessageVT() proto.Message {
+	return m.CloneVT()
+}
+
+func (m *DeviceModel) CloneVT() *DeviceModel {
+	if m == nil {
+		return (*DeviceModel)(nil)
+	}
+	r := &DeviceModel{
+		Name:       m.Name,
+		Identifier: m.Identifier,
+	}
+	if len(m.unknownFields) > 0 {
+		r.unknownFields = make([]byte, len(m.unknownFields))
+		copy(r.unknownFields, m.unknownFields)
+	}
+	return r
+}
+
+func (m *DeviceModel) CloneMessageVT() proto.Message {
+	return m.CloneVT()
+}
 
 func (m *Device) MarshalVT() (dAtA []byte, err error) {
 	if m == nil {