--- a/model/modelpb/headers_vtproto.pb.go
+++ b/model/modelpb/headers_vtproto.pb.go
@@ -25,6 +25,7 @@ import (
 	fmt "fmt"
 	io "io"
 
+	proto "google.golang.org/protobuf/proto"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 )
 
@@ -34,6 +35,29 @@ const (
 	// Verify that runtime/protoimpl is sufficiently up-to-date.
 	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
 )
+
+func (m *HTTPHeader) CloneVT() *HTTPHeader {
+	if m == nil {
+		return (*HTTPHeader)(nil)
+	}
+	r := &HTTPHeader{
+		Key: m.Key,
+	}
+	if rhs := m.Value; rhs != nil {
+		tmpContainer := make([]string, len(rhs))
+		copy(tmpContainer, rhs)
+		r.Value = tmpContainer
+	}
+	if len(m.unknownFields) > 0 {
+		r.unknownFields = make([]byte, len(m.unknownFields))
+		copy(r.unknownFields, m.unknownFields)
+	}
+	return r
+}
+
+func (m *HTTPHeader) CloneMessageVT() proto.Message {
+	return m.CloneVT()
+}
 
 func (m *HTTPHeader) MarshalVT() (dAtA []byte, err error) {
 	if m == nil {