--- a/model/modelpb/host_vtproto.pb.go
+++ b/model/modelpb/host_vtproto.pb.go
@@ -25,6 +25,7 @@ import (
 	fmt "fmt"
 	io "io"
 
+	proto "google.golang.org/protobuf/proto"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 )
 
@@ -34,6 +35,34 @@ const (
 	// Verify that runtime/protoimpl is sufficiently up-to-date.
 	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
 )
+
+func (m *Host) CloneVT() *Host {
+	if m == nil {
+		return (*Host)(nil)
+	}
+	r := &Host{
+		Os:           m.Os.CloneVT(),
+		Hostname:     m.Hostname,
+		Name:         m.Name,
+		Id:           m.Id,
+		Architecture: m.Architecture,
+		Type:         m.Type,
+	}
+	if rhs := m.Ip; rhs != nil {
+		tmpContainer := make([]string, len(rhs))
+		copy(tmpContainer, rhs)
+		r.Ip = tmpContainer
+	}
+	if len(m.unknownFields) > 0 {
+		r.unknownFields = make([]byte, len(m.unknownFields))
+		copy(r.unknownFields, m.unknownFields)
+	}
+	return r
+}
+
+func (m *Host) CloneMessageVT() proto.Message {
+	return m.CloneVT()
+}
 
 func (m *Host) MarshalVT() (dAtA []byte, err error) {
 	if m == nil {