--- a/model/modelpb/kubernetes_vtproto.pb.go
+++ b/model/modelpb/kubernetes_vtproto.pb.go
@@ -25,6 +25,7 @@ import (
 	fmt "fmt"
 	io "io"
 
+	proto "google.golang.org/protobuf/proto"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 )
 
@@ -34,6 +35,27 @@ const (
 	// Verify that runtime/protoimpl is sufficiently up-to-date.
 	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
 )
+
+func (m *Kubernetes) CloneVT() *Kubernetes {
+	if m == nil {
+		return (*Kubernetes)(nil)
+	}
+	r := &Kubernetes{
+		Namespace: m.Namespace,
+		NodeName:  m.NodeName,
+		PodName:   m.PodName,
+		PodUid:    m.PodUid,
+	}
+	if len(m.unknownFields) > 0 {
+		r.unknownFields = make([]byte, len(m.unknownFields))
+		copy(r.unknownFields, m.unknownFields)
+	}
+	return r
+}
+
+func (m *Kubernetes) CloneMessageVT() proto.Message {
+	return m.CloneVT()
+}
 
 func (m *Kubernetes) MarshalVT() (dAtA []byte, err error) {
 	if m == nil {